Extract documentation index modules
